requirements/base.txt
Django==1.11.6
django-cors-headers==2.1.0
django-environ==0.4.4
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
pytz==2017.2
six==1.11.0
